Updated platformio.ini check_flags and build_flags and improved compliance.

+++ lib/StabilizedVehicle/src/AHRS.h
@@ -89,15 +89,15 @@ private:
     uint32_t _imuDataReadyCount {0}; //<! data ready count, used in interrupt service routine
 
     // instrumentation member data
-    size_t _fifoCount {0};
+    uint32_t _fifoCount {0};
     uint32_t _timeCheck0 {0};
     uint32_t _timeChecksMicroSeconds[TIME_CHECKS_COUNT + 1] {};
 #if defined(AHRS_RECORD_TIMES_CHECKS)
-    inline void TIME_CHECK(size_t index) { _timeChecksMicroSeconds[index] = micros(); }
-    inline void TIME_CHECK(size_t index, uint32_t timeMicroSeconds) { _timeChecksMicroSeconds[index] = timeMicroSeconds; }
+    inline void TIME_CHECK(uint32_t index) { _timeChecksMicroSeconds[index] = micros(); }
+    inline void TIME_CHECK(uint32_t index, uint32_t timeMicroSeconds) { _timeChecksMicroSeconds[index] = timeMicroSeconds; }
 #else
-    inline void TIME_CHECK(size_t index) { (void)index;}
-    inline void TIME_CHECK(size_t index, uint32_t timeMicroSeconds) { (void)index; (void)timeMicroSeconds; }
+    inline void TIME_CHECK(uint32_t index) { (void)index;}
+    inline void TIME_CHECK(uint32_t index, uint32_t timeMicroSeconds) { (void)index; (void)timeMicroSeconds; }
 #endif
 
     // data synchronization primitives
